+++ lib/features/reports/presentation/screens/report_master.dart
@@ -205,7 +205,7 @@ class _ReportMasterState extends State<ReportMaster> {
                                 amount: TFormatter.formatToRupiah(
                                   report.averageSales == "NaN"
                                       ? 0
-                                      : int.parse(
+                                      : double.parse(
                                           report.averageSales,
                                         ),
                                 ),
@@ -235,7 +235,7 @@ class _ReportMasterState extends State<ReportMaster> {
                                 ),
                                 label: "Total Keuntungan",
                                 amount: TFormatter.formatToRupiah(
-                                  int.parse(
+                                  double.parse(
                                     report.totalProfit,
                                   ),
                                 ),
